// LZEventApp/src/app/chat-app/graphql/operation-result-types.ts
export type getAllEventsQuery = {
  getEvents: Array<{
    _typename: "Events",
    Id: int,
    EventName: string | null,
    EventId:int
  } | null> | null
}